requirements.txt
Django==2.1.4
django-hitcount==1.3.0
freeze==1.0.10
Pillow==5.3.0
pytz==2018.7
six==1.12.0
